+++ HARDWARE/PARAMETES/parameters.h
@@ -1,3 +1,11 @@
+/*
+ * @fileName: 
+ * @Author: Jone
+ * @LastEditors: Jone
+ * @description: page description
+ * @Date: 2023-08-30 10:08:09
+ * @LastEditTime: 2023-08-30 11:46:16
+ */
 #ifndef _PARAMETERS_H
 #define _PARAMETERS_H
 
@@ -29,7 +37,7 @@ extern float chuli_current_val;
 extern u8 ChuliDispflag;
 extern float chulisensorvalue;
 extern u8 motogear;
-extern u16 presetnpressure;
+extern float presetnpressure;
 extern u16 flowsetvalue;
 extern u16 pressuresetvalue;
 extern u8 speedupflag;
